requirements.txt
aiogram==3.15.0
telethon==1.36.0
python-dotenv==1.0.1

